MAJ SQL

+++ install/sql/filldb.sql
@@ -118,8 +118,8 @@ CREATE TABLE `radionuclide` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
   `code` varchar(8) NOT NULL,
   `name` varchar(255) NOT NULL,
-  `period` int(11) NOT NULL,
-  `coefficient` int(11) NOT NULL,
+  `period` float NOT NULL,
+  `coefficient` float NOT NULL,
   PRIMARY KEY (`id`)
 ) ENGINE=InnoDB AUTO_INCREMENT=1 DEFAULT CHARSET=utf8;
